UPSTREAM: mwifiex: pcie: use posted write to wake up firmware
Depending on system factors (e.g., the PCIe link PM state), the first
read to wake up the Wifi firmware can take a long time. There is no
reason to use a (blocking, non-posted) read at this point, so let's just
use a write instead. Write vs. read doesn't matter functionality-wise --
it's just a dummy operation. But let's make sure to re-write with the
correct "ready" signature, since we check for that in other parts of the
driver.

This has been shown to decrease the time spent blocking in this function
on RK3399.

BUG=chrome-os-partner:60714
TEST=Kevin wifi; also instrument this driver to check for delays

Change-Id: I041685313f8c4762feb89b0c5d08f188d9812219
Signed-off-by: Brian Norris <briannorris@chromium.org>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 062e008a6e83)