ARCv2: smp: MCIP: remove debug aid to halt all cores when one halts
This was really usefull when doing initial bringup and also for
occassional debug of software and hardware bugs. However in the new
smp-boot regime, non masters will "self" halt even for run-on-reset
configs. This will misinteract with the debug feature as in even master
will get halted when the non masters self halt.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>