rcu: Make call_rcu() do deferred NOCB wakeups
If a CONFIG_RCU_NOCB_CPUS kernel invokes call_rcu() with interrupts
disabled, wakeups must be deferred in order to avoid self-deadlock in the
cases where the disabled interrupts are due to scheduler locks being held.
In this case, a flag is set and is checked on entry to extended quiescent
states (usermode, idle), on exit from the RCU_SOFTIRQ handler, and when
the CPU in question goes offline.  However, a given CPU could avoid all
of those states for a considerable length of time.

This commit therefore allows a subsequent invocation of call_rcu() with
interrupts enabled to do the wakeup.  It also makes the wakeup function
clear the deferred-wakeup flag.

Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>