crypto: msm: qce50: fix timer setup race condition
Deactivation of the timer used to check on outstanding crypto requests in
bunch mode must be protected, otherwise a race condition may occur.  As a
result of the race condition, the timer may not be active to check if a
crypto request completed. If this occurs, the crypto completion handlers
will not be invoked.  To the upper layer, the requests appear as not
completing which can result in a process hang.

Change-Id: I130743f66eb40901c28c9ae8bbf18bd5117649f0
Acked-by: Che-Min Hsieh <cheminh@qti.qualcomm.com>
Signed-off-by: Yasir Malik <ymalik@codeaurora.org>