lib/find_bit.c: micro-optimise find_next_*_bit
This saves 32 bytes on my x86-64 build, mostly due to alignment
considerations and sharing more code between find_next_bit and
find_next_zero_bit, but it does save a couple of instructions.

There's really two parts to this commit.  First, the first half of the
test: (!nbits || start >= nbits) is trivially a subset of the second half,
since nbits and start are both unsigned.  Second, while looking at the
disassembly, I noticed that GCC was predicting the branch taken.  Since
this is a failure case, it's clearly the less likely of the two branches,
so add an unlikely() to override GCC's heuristics.

Link: http://lkml.kernel.org/r/1483709016-1834-1-git-send-email-mawilcox@linuxonhyperv.com
Signed-off-by: Matthew Wilcox <mawilcox@microsoft.com>
Acked-by: Yury Norov <ynorov@caviumnetworks.com>
Acked-by: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>