CHROMIUM: compat jail-control interface
We need a compat_ioctl function for 32-bit programs on 64-bit
kernels. In addition, since the struct size is different, we need
to define a compat struct and compat ioctl number for the
JAIL_CONTROL_ADD_DEVICE ioctl.

Note to future cherry-pickers: this can be squashed into the
previous device_jail commit to reduce the amount of patches we
need to carry forward.

BUG=chromium:644338
TEST=device_jail_utility and device_jail_fs on kevin

Change-Id: I8d6be07d170046ee0832a9b7f1dc9318b099ad3d
Reviewed-on: https://chromium-review.googlesource.com/437913
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>