target: Simplify and generalize the SPC-2 reservation implementation
Since the information in dev_res_bin_isid, dev_reservation_flags
and dev_reserved_node_acl overlaps with the information in
reserved_by, keep the latter and remove the former. An additional
benefit of this change is that RESERVE / RELEASE functionality is
extended to other protocols than iSCSI without having to modify
these protocols to add a sess_get_initiator_sid() method.

Signed-off-by: Bart Van Assche <bart.vanassche@sandisk.com>
Cc: Hannes Reinecke <hare@suse.com>
Cc: Christoph Hellwig <hch@lst.de>
Cc: David Disseldorp <ddiss@suse.de>