scatterlist: don't overflow length field
When called with a region of contiguous pages totaling > 4 GB of memory,
sg_alloc_table_from_pages() will overflow the length field, leading to a
corrupt scatter list. Fix this by tracking the number of pages we've
merged and start a new chunk when we would overflow.

Tested by building various page lists with contiguous 8GB regions and
observing that they are correctly split without overflowing length.

This isn't from normal read/write IO -- some applications want to
access large amounts of userspace memory directly from hardware, and it
is cleaner for them to manage one mapping than multiple 1GB or 2GB
mappings -- assuming the hardware can even support multiple mappings.
If they have room in their container to allocate and pin the memory,
we'd like to allow it.

There's definitely potential for problems downstream, even without
going through the filesystems and block layers -- we noticed this
potential issue while tracking down an bug in the IOMMU code when an
entry in the list was over 1GB.  We still see a benefit from building
the large entries, though -- it allows superpages in the IOMMU mapping
which helps the IOTLB cache.

We currently use sg_alloc_table_from_pages() to build the scatterlist
for dma_map_sg() but we could do it ourselves if you'd rather add a
length limit to the more general code.

Link: http://lkml.kernel.org/r/20170201212917.11278-1-dillow@google.com
Signed-off-by: David Dillow <dillow@google.com>
Cc: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>