regulator: qpnp: Modify the range selection logic
The range selection logic in select_voltage function finds the range
in which the new voltage request is inside of. Modify the range
selection logic to check the new voltage against min and max allowed
voltages of each range from bottom to top.

Also, modify the range arrays that doesn't follow the below condition:
	range[i].set_point_max_uV < range[i+1].set_point_min_uV

CRs-Fixed: 2002783
Change-Id: I1d9dfde824ad76f46e008988bc41b58320957518
Signed-off-by: Tirupathi Reddy <tirupath@codeaurora.org>