UPSTREAM: drm: mali-dp: Remove mclk rate management
The rate of mclk depends on the use-case. If no downscaling is required,
then mclk == pxlclk is a valid option; with downscaling however, the
rate at which mclk runs determines how much a plane can be downscaled
before composition. This is a system integration + power management
issue that is more suited to firmware rather than this driver.

BUG=chromium:707244
TEST=Build and run

Change-Id: I8030ee4283749b3d4fba166d25dc343467fa88d6
Signed-off-by: Mihail Atanassov <mihail.atanassov@arm.com>
Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 9a8b0a230aca)