UPSTREAM: PCI: rockchip: Set vendor ID from local core config space
The TRM says the vendor ID in the RC's configure space can be rewritten
and the value must be the same as the value read from the local core
configure space.  But we misread that and didn't notice it before.  Actually
we should only able to rewrite it from the local core configure space.

Fix that issue to make lspci show the correct IP vendor infomation.

BUG=chromium:707244
TEST=Build and run

Change-Id: If7078f4f1c54382d4f8b514bda4385386a047e00
Signed-off-by: Shawn Lin <shawn.lin@rock-chips.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 5800790a925b)