mm, vmscan: clear PGDAT_WRITEBACK when zone is balanced
Hillf Danton pointed out that since commit 1d82de618dd ("mm, vmscan: make
kswapd reclaim in terms of nodes") that PGDAT_WRITEBACK is no longer
cleared.  It was not noticed as triggering it requires pages under
writeback to cycle twice through the LRU and before kswapd gets stalled.
Historically, such issues tended to occur on small machines writing
heavily to slow storage such as a USB stick.  Once kswapd stalls, direct
reclaim stalls may be higher but due to the fact that memory pressure is
required, it would not be very noticable.  Michal Hocko suggested removing
the flag entirely but the conservative fix is to restore the intended
PGDAT_WRITEBACK behaviour and clear the flag when a suitable zone is
balanced.

Fixes: 1d82de618ddd ("mm, vmscan: make kswapd reclaim in terms of nodes")
Link: http://lkml.kernel.org/r/20170203203222.gq7hk66yc36lpgtb@suse.de
Signed-off-by: Mel Gorman <mgorman@suse.de>
Acked-by: Johannes Weiner <hannes@cmpxchg.org>
Acked-by: Michal Hocko <mhocko@suse.com>
Acked-by: Hillf Danton <hillf.zj@alibaba-inc.com>
Cc: Minchan Kim <minchan.kim@gmail.com>
Cc: Rik van Riel <riel@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>