mm: do not access page->mapping directly on page_endio
With rw_page, page_endio is used for completing IO on a page and it
propagates write error to the address space if the IO fails.  The problem
is it accesses page->mapping directly which might be okay for file-backed
pages but it shouldn't for anonymous page.  Otherwise, it can corrupt one
of field from anon_vma under us and system goes panic randomly.

Link: http://lkml.kernel.org/r/1487741964-17913-1-git-send-email-minchan@kernel.org
Signed-off-by: Minchan Kim <minchan@kernel.org>
Cc: Matthew Wilcox <willy@infradead.org>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>